+++ src/assets/mockGuru/index.ts
@@ -1,0 +1,3 @@
+export default {
+  imageContent: require('./imageContent.png'),
+};